nat: replace readyCtx with just a channel
Context are usually used to establish a call function hierarchy, and so
many functions take contexts as arguments. Since this part of the code
has several of them, it can get hard to figure out exactly which one to
use and when.

This commit replaces `readyCtx` with a simple channel since its main
function is to unblock a goroutine. Being a channel removes the
expectations that it may be passed in a function call, making its scope
more clear.

It also removes the `defer` statement since we don't want to unblock the
happy path in case of an error. The new channel is guaranteed to only be
closed when everything succeeds.